[Feat]: 땅따먹기 게임 프로필 생성
땅따먹기 인게임 프로필 생성
플레이어 닉네임, 점수가 포함되어 있음
점수 -> 내가 먹은 땅의 개수
