+++ Assets/LeeYunJeong/Scripts/TPS_Scripts/TPSPlayerProfileManager4.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86f43f90f10a61943b8c7839b13c8df3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
