+++ Assets/LeeYunJeong/Scripts/Ttangttameokgi/PlayerProfileManager4.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1782cd8818ff53c4eb64068967f7c445
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
